Add Create, searchById, Delete function for plans

+++ backend/src/main/java/com/telecom/repository/PlanRepository.java
@@ -6,7 +6,7 @@ import com.telecom.model.PhonePlan;
 
 // class name and primary key type
 
-public interface PlanRepository extends JpaRepository<PhonePlan, Integer> {
+public interface PlanRepository extends JpaRepository<PhonePlan, Long> {
 
 	// tell JPA to use a select with "findBY"
 	PhonePlan findByPlanName(String planName);
